admin side all fangtionality completed and backend some parts are changed

+++ server/controllers/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/category')
+// const fs = require('fs');
 
 const getCategory = async (req, res) => {
   try {
@@ -9,7 +10,19 @@ const getCategory = async (req, res) => {
   }
 };
 
+
+const getCategoryById = async (req, res) => {
+  try {
+    const data = await Category.findOne({ _id: req.params.id });
+    res.status(200).json({ data, message: 'product found successfully' });
+  } catch (error) {
+    console.log(error.message);
+    res.status(400).json({ message: error?.message ?? "Something went wrong !" });
+  }
+}
+
 const addCategory = async (req, res) => {
+  console.log('hello');
   const { name, desc } = req?.body
   const image = req?.file?.filename
   try {
@@ -31,21 +44,52 @@ const addCategory = async (req, res) => {
     console.log(error);
   }
 }
-
-const deleteCategory = async (req, res) => {
+const updateCategory = async (req, res) => {
+  const {_id,name, desc ,isAvailable} = req?.body
+  const image = req?.file?.filename
+  console.log('hello2',image);
   try {
-    const id = req.query.id
-    const data = await Category.deleteOne({ _id: id });
-    fs.unlink(`public/uploads/${data?.image}`, (err) => {
-      if (err) {
-        console.error('Error deleting image:', err);
-        return;
-      }
-      console.log('Image deleted successfully.');
-    });
-    res.status(200).json({ message: 'category deleted successfully' });
+    const data = await Category.findById(_id);
+    if (!data) {
+      return res.status(404).json({ message: 'Category not found' });
+    }
+    if (image) {
+      fs.unlink(`public/uploads/${data?.image}`, (err) => {
+        if (err) {
+          console.error('Error deleting image:', err);
+          return;
+        }
+        console.log('Image deleted successfully.');
+      });
+    }
+    await Category.updateOne({ _id }, {
+      $set: { name, desc,isAvailable, ...(image && { image }) }
+    })
+    res.status(200).json({ data, message: 'Category updated successfully' });
   } catch (error) {
     console.log(error);
+    return res.status(500).json({ message: error?.message ?? 'Something went wrong' })
+  }
+}
+
+const deleteCategory = async (req, res) => {
+  const { id } = req.params;
+  try {
+    const data = await Category.findByIdAndDelete(id);
+    if (!data) {
+      return res.status(404).json({ message: 'Category not found' });
+    }
+    // fs.unlink(`public/uploads/${data?.image}`, (err) => {
+    //   if (err) {
+    //     console.error('Error deleting image:', err);
+    //     return;
+    //   }
+    //   console.log('Image deleted successfully.');
+    // });
+    res.status(200).json({ message: 'Category deleted successfully' });
+  } catch (error) {
+    console.log(error);
+    return res.status(500).json({ message: error?.message ?? 'Something went wrong' })
   }
 }
 
@@ -53,4 +97,6 @@ module.exports = {
     getCategory,
     addCategory,
     deleteCategory,
+    updateCategory,
+    getCategoryById
   }
